requirements.txt
pandas
numpy
PyPDF2
openpyxl
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
python-dotenv
flask
flask-cors
schedule
